init script and virtualenv
I had to make a view changes to the init script for it to run on bash, hopefully this does not
break any backwards compatability on your end.

I added the option to install virtualenv, as otherwise you have to re-do the script if you
do not have it installed (as I did not).